feat: add unique adapter
