+++ packages/ouro-core/src/adapter/index.js
@@ -11,4 +11,5 @@ export { default as Skip } from './skip'
 export { default as TakeWhile } from './take-while'
 export { default as Take } from './take'
 export { default as Tap } from './tap'
+export { default as Unique } from './unique'
 export { default as Zip } from './zip'
